Some changes to the exercises repo:
1. Added READMEs that are taken from the latest training instructions (HLRS).
Some have been added to the associated directory of exercises. Others (such as Rocprof
and Rocgdb) have been placed in a dir README_DIR that collects READMEs that are not
currentl directly associated with exercises in the repo but that we should still keep track of.

2. Added REPO_DIR environment variable to the test scripts
so that it is not expected anymore that the repo is cloned in the home directory.

3. Started a top level README.md file where we will provide information on the repo
content structure. Might have to do a per-task sorting rather than a per-topic.